VolumePanel: Don't timeout quickly if expanded
The volume panel disappears after only 1.5 seconds if the active stream is
media. This check currently takes precedence over the state of the panel
(whether it is expanded). Setting volume keys to control media volume (not
ringtone volume) forces the active stream to always report media, making the
panel disappear very quickly whenever this setting is enabled.

Move the media check to after the expanded state check so the panel doesn't
timeout prematurely if it is expanded.

Change-Id: If295bc04e5903c29e29bf9f6b55b3e5a1ab5f3bd
Signed-off-by: AudioGod <adi.nth@outlook.com>