Chrome floating fix:
Chrome uses the same address for
FLAG_ACTIVITY_RETAIN_IN_RECENTS = 0x00002000
and the porter didn't took this into account and added
the FLAG_FLOATING_WINDOW intent the same adress and
that is why Chrome was always starting in floating mode.

Side effect of this change is that non regular Chrome starts
always in full screen (normal mode), but if you select in
Chrome settings to "Merge tabs and apps" it will act as other
apps. Starting in float when float is invoked from Slim recents.

Change-Id: I5f3f2a9ba86d8e1fd30d2d4cbc7d5667d6549cb5
Signed-off-by: AudioGod <adi.nth@outlook.com>
Signed-off-by: AudioGod <audiogod@sonic-developers.com>