Authenticate When Keyguard Resumes From Camera
Before this patch, fingerprint would stop working if a user went
to the camera inside keyguard. This is because keyguard would
not re-authenticate when it resumed from camera.

This patch introduces lazy authentication because Camera's onPause
can happen before or concurrently to Keyguard's occlusion callback.
This means Keyguard can get in a situation where it tries to
authenticate when FingerprintService is already authenticating.

Issue-Id: SAMBAR-205

Change-Id: I57bdad5140d172c0ffb963010e12cb72e85663e7
Signed-off-by: AudioGod <AudioGod@Sonic-Developers.com>